Import types, move constants
