+++ src/commands/game_commands/skip.ts
@@ -13,11 +13,11 @@ import { GuildTextableMessage } from "../../types";
 import { KmqImages } from "../../constants";
 import MessageContext from "../../structures/message_context";
 import CommandPrechecks from "../../command_prechecks";
-import EliminationScoreboard from "../../structures/elimination_scoreboard";
+import type EliminationScoreboard from "../../structures/elimination_scoreboard";
 import { state } from "../../kmq_worker";
-import Round from "../../structures/round";
+import type Round from "../../structures/round";
 import Session from "../../structures/session";
-import GuildPreference from "../../structures/guild_preference";
+import type GuildPreference from "../../structures/guild_preference";
 import CommandArgs from "../../interfaces/command_args";
 import HelpDocumentation from "../../interfaces/help";
 import { GameType } from "../../enums/game_type";
