+++ src/commands/interfaces/base_command.ts
@@ -2,7 +2,7 @@ import CallFunc from "../../interfaces/call_func";
 import CommandValidations from "../../interfaces/command_validations";
 import HelpDocumentation from "../../interfaces/help";
 import PrecheckArgs from "../../interfaces/precheck_args";
-import GuildPreference from "../../structures/guild_preference";
+import type GuildPreference from "../../structures/guild_preference";
 
 export default interface BaseCommand {
     call: CallFunc;
