+++ src/constants.ts
@@ -8,6 +8,7 @@ import { OstPreference } from "./enums/option_types/ost_preference";
 import { ReleaseType } from "./enums/option_types/release_type";
 import { SeekType } from "./enums/option_types/seek_type";
 import { ShuffleType } from "./enums/option_types/shuffle_type";
+import { SpecialType } from "./enums/option_types/special_type";
 import { SubunitsPreference } from "./enums/option_types/subunit_preference";
 
 // eslint-disable-next-line import/prefer-default-export
@@ -41,3 +42,40 @@ export const DEFAULT_SEEK = SeekType.RANDOM;
 export const DEFAULT_GENDER = [Gender.FEMALE, Gender.MALE, Gender.COED];
 export const DEFAULT_LIMIT = 500;
 export const SONG_START_DELAY = 3000;
+
+export const specialFfmpegArgs = {
+    [SpecialType.REVERSE]: (seek: number) => ({
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "areverse"],
+    }),
+    [SpecialType.SLOW]: (seek: number) => ({
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "rubberband=tempo=0.5"],
+    }),
+    [SpecialType.FAST]: (seek: number) => ({
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "rubberband=tempo=1.5"],
+    }),
+    [SpecialType.FASTER]: (seek: number) => ({
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "rubberband=tempo=2"],
+    }),
+    [SpecialType.LOW_PITCH]: (seek: number) => ({
+        // 3 semitones lower
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "rubberband=pitch=0.840896"],
+    }),
+    [SpecialType.HIGH_PITCH]: (seek: number) => ({
+        // 4 semitones higher
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "rubberband=pitch=1.25992"],
+    }),
+    [SpecialType.NIGHTCORE]: (seek: number) => ({
+        inputArgs: ["-ss", seek.toString()],
+        encoderArgs: ["-af", "rubberband=pitch=1.25992:tempo=1.25"],
+    }),
+};
+
+// z = chinese, j = japanese, e = english, s = spanish
+export const FOREIGN_LANGUAGE_TAGS = ["z", "j", "e", "s"];
+export const NON_OFFICIAL_VIDEO_TAGS = ["c", "d", "a", "r", "v", "x", "p"];
