+++ src/kmq_worker.ts
@@ -15,12 +15,12 @@ import {
 import BotListingManager from "./helpers/bot_listing_manager";
 import RateLimiter from "./rate_limiter";
 import dbContext from "./database_context";
-import KmqClient from "./kmq_client";
+import type KmqClient from "./kmq_client";
 import ReloadCommand from "./commands/admin/reload";
 import EvalCommand from "./commands/admin/eval";
 import LocalizationManager from "./helpers/localization_manager";
 import Session from "./structures/session";
-import State from "./interfaces/state";
+import type State from "./interfaces/state";
 import { EnvType } from "./enums/env_type";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
