+++ src/structures/elimination_scoreboard.ts
@@ -1,5 +1,5 @@
 import Scoreboard from "./scoreboard";
-import EliminationPlayer from "./elimination_player";
+import type EliminationPlayer from "./elimination_player";
 import SuccessfulGuessResult from "../interfaces/success_guess_result";
 
 export const DEFAULT_LIVES = 10;
